feat: 修改chatSearchContent、chatSuggestionContent、chatList的定义位置；删除部分类型声明；添加chatbot和chat-action的类型导出
