+++ packages/pro-components/chat/chatbot/index.ts
@@ -2,7 +2,6 @@ import 'tdesign-web-components/lib/chatbot';
 import type { DefineComponent } from 'vue';
 import { omiVueify } from 'omi-vueify';
 import type {
-  TdChatbotApi,
   TdChatListProps,
   TdChatProps,
   TdChatSearchContentProps,
@@ -13,4 +12,15 @@ export const ChatBot = omiVueify('t-chatbot', {
   methodNames: ['addPrompt', 'regenerate', 'selectFile', 'registerMergeStrategy'],
 }) as DefineComponent<TdChatProps>;
 
+export const ChatSearchContentComponent = omiVueify('t-chat-search-content', {
+  methodNames: [],
+}) as DefineComponent<TdChatSearchContentProps>;
+
+export const ChatSuggestionContentComponent = omiVueify('t-chat-suggestion-content', {
+  methodNames: [],
+}) as DefineComponent<TdChatSuggestionContentProps>;
+
+export const ChatListComponent = omiVueify('t-chat-list', {
+  methodNames: ['scrollToBottom'],
+}) as DefineComponent<TdChatListProps>;
 export default ChatBot;
