+++ packages/pro-components/chat/type.ts
@@ -421,3 +421,5 @@ export type * from 'tdesign-web-components/lib/chat-sender/type';
 export type * from 'tdesign-web-components/lib/filecard/type';
 export type * from 'tdesign-web-components/lib/chat-message/index';
 export type * from 'tdesign-web-components/lib/chatbot/type';
+export * from 'tdesign-web-components/lib/chatbot/core/utils';
+export * from 'tdesign-web-components/lib/chat-action/type';
